requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.1
asyncpg>=0.27.0
supabase>=1.0.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
pytz>=2023.3
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pytest>=7.3.1
httpx>=0.24.1
